pyproject.toml
[tool.poetry]
name = "fastapi_todo"
version = "0.1.0"
description = ""
authors = ["sharofiddin <cyberpancher10@gmail>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
Jinja2 = "^3.1.2"
SQLAlchemy = "^2.0.4"
python-dotenv = "^0.21.1"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
